+++ apps/admin/src/views/tools/MenuEditor/menuEditor.jsx
@@ -86,6 +86,7 @@ class ToolOptions extends Component {
     draggingEnabled: false,
     searchImplemented: true,
     enablePrint: true,
+    directPrint: false,
     pdfLinks: [{ name: "", link: "" }],
     closePanelOnMapLinkOpen: false,
     displayLoadingOnMapLinkOpen: false,
@@ -113,19 +114,22 @@ class ToolOptions extends Component {
     this.type = "documenthandler";
     this.mapSettingsModel = props.model;
     this.menuEditorModel = this.getMenuEditorModel();
-    this.menuEditorModel.listAllAvailableDocuments(this.state.folder).then((list) => {
-      this.setState({
-        availableDocuments: list,
-      })
-    });
-    
-    this.useDocumentFolders = this.menuEditorModel.config.use_document_folders ?? false;
+    this.menuEditorModel
+      .listAllAvailableDocuments(this.state.folder)
+      .then((list) => {
+        this.setState({
+          availableDocuments: list,
+        });
+      });
+
+    this.useDocumentFolders =
+      this.menuEditorModel.config.use_document_folders ?? false;
 
     if (this.useDocumentFolders) {
       this.menuEditorModel.loadFolders().then((list) => {
         this.setState({
           folders: list,
-        })
+        });
       });
     }
   }
@@ -143,7 +147,7 @@ class ToolOptions extends Component {
               this.updateTreeValidation(this.state.tree);
               // I hate myself... This should be avoided at all costs!!!
               this.forceUpdate();
-            }
+            },
           );
         });
     });
@@ -196,6 +200,7 @@ class ToolOptions extends Component {
         draggingEnabled: tool.options.draggingEnabled || false,
         searchImplemented: tool.options.searchImplemented,
         enablePrint: tool.options.enablePrint,
+        directPrint: tool.options.directPrint || false,
         pdfLinks: tool.options.pdfLinks || [{ name: "", link: "" }],
         closePanelOnMapLinkOpen: tool.options.closePanelOnMapLinkOpen,
         displayLoadingOnMapLinkOpen:
@@ -269,7 +274,7 @@ class ToolOptions extends Component {
   saveFromMenuEditor() {
     this.menuConfig = this.menuEditorModel.exportTreeAsMenuJson(
       this.state.tree,
-      this.menuConfig
+      this.menuConfig,
     );
     this.setState({ menuConfig: this.menuConfig }, () => {
       this.save();
@@ -291,6 +296,7 @@ class ToolOptions extends Component {
         height: this.state.height,
         searchImplemented: this.state.searchImplemented,
         enablePrint: this.state.enablePrint,
+        directPrint: this.state.directPrint,
         pdfLinks: this.state.pdfLinks,
         closePanelOnMapLinkOpen: this.state.closePanelOnMapLinkOpen,
         displayLoadingOnMapLinkOpen: this.state.displayLoadingOnMapLinkOpen,
@@ -316,7 +322,7 @@ class ToolOptions extends Component {
             alert: true,
             alertMessage: "Uppdateringen lyckades",
           });
-        }
+        },
       );
     }
     if (!this.state.active) {
@@ -440,7 +446,7 @@ class ToolOptions extends Component {
   addHeaderRowToTreeStructure = (treeData) => {
     treeData.unshift({
       title: this.getHeader(
-        this.menuEditorModel.canSave(this.getTreeWithoutHeader(treeData))
+        this.menuEditorModel.canSave(this.getTreeWithoutHeader(treeData)),
       ),
       disabled: true,
       children: [],
@@ -509,7 +515,7 @@ class ToolOptions extends Component {
     treeNode.title = this.getRowTitleComponent(
       treeNode.menuItem,
       treeNode.children,
-      treeNode.key
+      treeNode.key,
     );
   };
 
@@ -568,21 +574,21 @@ class ToolOptions extends Component {
             newTree,
             foundDragNode,
             foundDropNode,
-            info
+            info,
           );
         } else {
           this.menuEditorModel.addToGap(
             newTree,
             foundDragNode,
             foundDropNode,
-            info
+            info,
           );
         }
       } else {
         this.menuEditorModel.addToDropNode(
           newTree,
           foundDragNode,
-          foundDropNode
+          foundDropNode,
         );
       }
 
@@ -600,7 +606,7 @@ class ToolOptions extends Component {
   saveNewTree = (newTree) => {
     this.updateTreeValidation(newTree);
     newTree[0].title = this.getHeader(
-      this.menuEditorModel.canSave(this.getTreeWithoutHeader(newTree))
+      this.menuEditorModel.canSave(this.getTreeWithoutHeader(newTree)),
     );
     this.setState({ tree: newTree });
   };
@@ -625,7 +631,7 @@ class ToolOptions extends Component {
       if (child.key === treeNodeToDelete.key) {
         nodeArrayToSearch.splice(
           nodeArrayToSearch.indexOf(treeNodeToDelete),
-          1
+          1,
         );
       }
     });
@@ -910,6 +916,23 @@ class ToolOptions extends Component {
             &nbsp;
             <label htmlFor="enablePrint">Utskrift aktiverad</label>
           </div>
+          {this.state.enablePrint && (
+            <div style={{ paddingLeft: "16px", borderLeft: "2px solid #ddd" }}>
+              <input
+                id="directPrint"
+                name="directPrint"
+                type="checkbox"
+                onChange={(e) => {
+                  this.handleInputChange(e);
+                }}
+                checked={this.state.directPrint}
+              />
+              &nbsp;
+              <label htmlFor="directPrint" style={{ width: "auto" }}>
+                Endast aktiv sida kommer skrivas ut (utan dokumentval)
+              </label>
+            </div>
+          )}
           <div>
             {this.state.pdfLinks &&
               this.state.pdfLinks.map((pdfLink, index) => (