+++ apps/client/src/models/layers/WMSLayer.jsx
@@ -121,6 +121,8 @@ class WMSLayer {
         ? config.visibleAtStartSubLayers
         : this.subLayers
     );
+    this.layer.set("hasLabelStyle", config.hasLabelStyle);
+    this.layer.set("showLabelLayer", false);
     this.layer.getSource().set("url", config.url);
     this.type = "wms";
     this.bindHandlers();